Move all tests for shallow fetching into their own module
Note that those which test both with and without shallow still go
into the shallow module.
